class to find frontier centriods added
this is not tested and cannot be run yet. However, the code is all written
and the logic should work. If it will be used as a class then a blank
constructer must be created as the class does not have any instance variables
it is just a collection of methods that all take in params for any information
they need

-
mike